pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "pybind11>=2.10.0", "cmake>=3.12", "ninja"]
build-backend = "setuptools.build_meta"

[project]
name = "slidescanner"
version = "0.1.0"
description = "Canon camera control application for slide scanning"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Coco", email = "g.colinharley@gmail.com"},
]
keywords = ["camera", "canon", "gtk", "photography", "scanning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: C++",
    "Topic :: Multimedia :: Graphics :: Capture :: Digital Camera",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "PyGObject>=3.42.0",
    "Pycairo>=1.21.0",
    "numpy>=1.21.0",
    "Pillow>=8.3.0",
    "exifread>=2.3.0",
    "piexif>=1.1.3",
    "matplotlib>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.0",
    "black>=21.0.0",
    "flake8>=3.9.0",
    "mypy>=0.910",
    "pybind11-stubgen>=0.10.0",
    "pyright>=1.1.300",
    "types-PyGObject",
]

[project.urls]
Homepage = "https://github.com/coco/slidescanner"
Repository = "https://github.com/coco/slidescanner.git"
Issues = "https://github.com/coco/slidescanner/issues"

[project.scripts]
slidescanner = "main:main"

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-dir]
"" = "."

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff.lint]
# Globally disable E402
ignore = ["E402"]
[tool.pyright]
# pyright config
reportMissingModuleSource = "none"


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
